got things working
So quick run through of what i have done. RIght now the onTouch only changes the background of the 2nd item in the list. Added a list view for the possible matches of applications. need to figure out how to get my gridview to minimize height of each node when the count gets over a certain value.