added more stuff
So i added in a class that will handle the coordinate system for each individual view. this will allow us to determine how to find out where we are touching inside the view. Will also allow us, at a future date, to determine if we are outside the bounds of the gridview if we wanted. changed a couple namespaces just so they made some more sense. also renamed NodeOfGridView to something mroe common IndividualNode.